app/requirements.txt
requests==2.27.1
fastapi==0.78.0
uvicorn==0.17.6
pytest==7.1.2
pandas==1.3.5
pyarrow==5.0.0
jinja2==3.1.2
sqlalchemy==1.4.37
psycopg2-binary==2.9.3